backend/requirements.txt
# Core
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0

# Database
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Settings & Validation
pydantic>=2.6.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# Auth & Security
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# AI / LLM / RAG
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.1.0
langchain-community>=0.3.0
pinecone>=6.0.0
arxiv>=2.1.0
google-search-results>=2.4.2
httpx>=0.27.0
